mycelium/spore/requirements.txt
asyncio-nats-client==0.11.4
nats-py==2.7.2
paho-mqtt==2.1.0
aiohttp==3.9.5
asyncpg==0.29.0
pymodbus==3.7.0
pydantic==2.7.1
pydantic-settings==2.3.0
structlog==24.1.0
httpx==0.27.0
uvicorn==0.30.0
python-ulid==2.2.0
